Add ipxe generation endpoint & persistence
The former allows you to generate an ipxe script and, if persistence
auth is turned on, save the resulting config for later. When saved, the
server will return a an ID for the ipxe script that can be passed to
the /boot/:id path. Additional cfg_* arguments can be passed to the
boot path to also configure the autoinstall.cfg.