Add HEAD support
This is needed for Packet.com support since their server deployment
will send an initial HEAD request against the iPXE URL. As-is, this is
actually more involved than is needed, since it's just checking at the
end if it should write a body. This is perfectly fine, just a little
weird.